requirements.txt
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.1
docopt==0.6.2
feedgen==1.0.0
firebase-admin==6.5.0
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.3
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
lxml==5.2.2
msgpack==1.0.8
num2words==0.5.13
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==2.2.2
uuid7==0.1.0